requirements.txt
praw==7.7.1
python-telegram-bot==13.15
requests==2.31.0
